requirements.txt
fastapi
uvicorn[standard]
python-dotenv
PyGithub
google-generativeai
requests